Move and rename pull_test from mod.rs to pull.rs
This test was testing the HTTP part of pull.

Towards #290
